migrate the height/width logic around the PDF page renderer from double to int. This should help improve cache performance a bit as now == equality comparisons should actually deliver instead of possibly maybe matching.
Also fixed a width edit I missed (the old cache and calc stuff only reckoned with height :-( -- https://github.com/jimmejardine/qiqqa-open-source/issues/9 )

Added cache retrieval code which picks a higher rez image when already present: this should ultimate reduce the pressure on the PDF renderer, of course depending on the order of incoming requests (thumbnails in side panel vs full rez main page views)

+++ Qiqqa/Documents/PDF/PDFAnnotationToImageRenderer.cs
@@ -7,7 +7,7 @@ namespace Qiqqa.Documents.PDF
 {
     internal class PDFAnnotationToImageRenderer
     {
-        public static Image RenderAnnotation(PDFDocument pdf_document, PDFAnnotation pdf_annotation, float dpi)
+        public static Image RenderAnnotation(PDFDocument pdf_document, PDFAnnotation pdf_annotation, int dpi)
         {
             using (MemoryStream ms = new MemoryStream(pdf_document.PDFRenderer.GetPageByDPIAsImage(pdf_annotation.Page, dpi)))
             {
